Fix Bug #2115 - CCCUSB::readDelays must be done as a read32.

+++ usb/ccusb/ccusb/CCCUSB.h
@@ -213,10 +213,10 @@ public:
     }
 
 
-    int readDelays(uint16_t& value);
+    int readDelays(uint32_t& value);
     int writeDelays(uint16_t value);
     unsigned  readDelays() {		/* swig */
-      uint16_t d;
+      uint32_t d;
       readDelays(d);
       return d;
     }
